ai-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.2
aiohttp==3.9.5
httpx==0.27.0
scikit-learn==1.5.0
python-dotenv==1.0.1
pydantic==2.7.1
